+++ src/App.tsx
@@ -1,10 +1,5 @@
 import Treege from "@/features/Treege";
 
-const App = () => (
-  <Treege
-    // endPoint="https://stg.api.tracktor.fr"
-    authToken="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJtaWNrYSIsImV4cCI6MTY2Njc5OTA4OH0._I7oMj70Taw2CO7QhsC4wREHtZmJ89w5_EpWby-zmR8"
-  />
-);
+const App = () => <Treege />;
 
 export default App;