+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Fix translation
 - Improve tree select
 - Disable some features when endpoint is not provided
+- Documentation